requirements.txt
# Requirements for the PDF processing and NLP API

# Web Framework to create the API
fastapi

# ASGI server to run FastAPI
uvicorn

# Required by FastAPI for file uploads
python-multipart

# For PDF text extraction
PyMuPDF

# For language detection
langdetect

# For making HTTP requests to external services
requests

# For NLP tasks
# Installation: python -m spacy download de_dep_news_trf
spacy

# Document Structuring Library (with all local PDF processing dependencies)
unstructured[local-inference]

# tesseract for OCR
pytesseract
# Installation: https://digi.bib.uni-mannheim.de/tesseract/
# Umgebungsvariable: TESSDATA_PREFIX = C:\Users\info\AppData\Local\Programs\Tesseract-OCR

# For advanced NLP tasks and models
transformers

# Sentencepiece is often required by tokenizers
sentencepiece

# PyTorch is the deep learning backend
torch

# For vector databases and similarity search
sentence-transformers 

# For managing and querying vector databases
chromadb

# For text splitting and chunking
langchain-text-splitters

# For dimensionality reduction and visualization
umap-learn

# For plotting and visualization
matplotlib 

# For numerical operations and data manipulation
numpy

# For JSON manipulation and validation
json_repair
